Move attendances mock tests / test data to new files to keep tests readable once more tests are added
