+++ tests/test_mock_api.py
@@ -1,5 +1,5 @@
 import re
-from datetime import date, timedelta
+from datetime import date
 from typing import Any, Dict
 
 import pytest
@@ -145,35 +145,6 @@ def test_get_absence_types():
     for source_dict, at in zip(json_dict_absence_types['data'], absence_types):
         target_dict = at.to_dict()
         assert source_dict == target_dict
-
-
-@responses.activate
-def test_get_attendance():
-    # mock the get absences endpoint (with different array offsets)
-    responses.add(
-        responses.GET, re.compile('https://api.personio.de/v1/company/attendances?.*offset=0.*'),
-        status=200, json=json_dict_attendance_rms, adding_headers={'Authorization': 'Bearer foo'})
-    responses.add(
-        responses.GET, re.compile('https://api.personio.de/v1/company/attendances?.*offset=3.*'),
-        status=200, json=json_dict_empty_response, adding_headers={'Authorization': 'Bearer bar'})
-    # configure personio & get absences for alan
-    personio = mock_personio()
-    attendances = personio.get_attendances(2116366)
-    # validate
-    assert len(attendances) == 3
-    selection = [a for a in attendances if "release" in a.comment.lower()]
-    assert len(selection) == 1
-    release = selection[0]
-    assert "free software" in release.comment
-    assert release.date == date(1985, 3, 20)
-    assert release.start_time == timedelta(seconds=11*60*60)
-    assert release.end_time == timedelta(seconds=12.5*60*60)
-    assert release.break_duration == 60
-    assert release.employee_id == 2116366
-    # validate serialization
-    source_dict = json_dict_attendance_rms['data'][0]
-    target_dict = release.to_dict()
-    compare_labeled_attributes(source_dict, target_dict)
 
 
 def mock_personio():
